Some supporting classes and a struct for the GMTK_Tie class.
The class ClusterableObject<T> is a wrapper for any parameter type <T>
that can be clustered and, when instantiated (e.g. ClusterableMean),
contains implementations of all the dissimilarity measures available
for type T.

The parent class Clusterable is to provide a single class type and
member function for the generic clustering routines in the GMTK_Tie
class - these routines do not need to know about the type T, since
they can call the function dissimilarityFunction() to request a
Clusterable object to measure its dissimilarity to another Clusterable
object. dissimilarityFunction is a pointer to a function, which is set
when, e.g. a ClusterableMean object is constructed.

The struct Cluster is just a simple holder for a list of Clusterable
objects, along with values for the cluster's current size and
occupancy

Finally, there are a few utility and convenience functions dealing
with looking up parameters and their subcomponents, just to make the
implementation simpler and prettier.