Fix a bug in the command line arguments checking that would lead to an EM bug if
  -cliqueTableNormalize 0.0 is given.

This bug didn't seem to manifest, but it would have been a serious EM training bug
had this occured. refs #167.

(note that the actual change was done in a previous check in and that had an accidental reference
to 168 but #167 is the right ticket for this change).  The actual change set is
changeset [a9b77c7f1e0a]