Skip test case by default; refs #572
This test requires GMTK to be built with HDF5 support.
Furthermore, GMTK must be compiled with the same C++
compiler as the HDF5 library. I haven't thought of a
way to test for that to automatically disable the test
when it's not applicable, so it is skipped unless the
{{{GMTK_ENABLE_HDF5_TESTS}}} environment variable is
non-empty.